+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added file name to the file template picker item's description to improve matching
 - Always show the icon to indicate the type of current testcase and make it clickable to toggle testcase interactivity
 - Also show the dropdown button (that does nothing) for compiling status to keep the toolbars consistent
+- Shorten the compiling error to just "CE" in stress tester for consistency
 
 ### Fixed